Python/Final project/requirements.txt
selenium
pytest
python-dotenv